Fix severe bug for cgp evolution that lead to homogeneous cgp populations
The genome in cgp individuals is stored in lists.
Since assignment in Python never copies
data (https://nedbatchelder.com/text/names.html) appending individuals
to parent and offspring lists lead to multiple individuals maintaining
references to the same genome. This bug severely decreased the
diversity in the population as after a while all individuals are
defined by the same genome. Cloning individuals when creating parent
and offspring lists prevents this, but requires each subclass of
AbstractPopulation to define how individuals can be cloned.