Avoid catching of SymPy expression
These kind of problems should not be caught but propagated to the user
who will know how to deal with them appropriately. This specific case
is even worse: since the InvalidSympyexpression is only defined
locally in _validate_sympy_expr, we can't even catch that specific
expression outside on the user level, forcing us to use a catch all.